back-end/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
openai==1.16.2
packaging==24.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
